+++ app/src/main/java/com/example/npod/data/dto/PictureOfTheDayDTO.kt
@@ -1,5 +1,6 @@
 package com.example.npod.data.dto
 
+import com.example.npod.domain.models.PictureOfTheDay
 import com.google.gson.annotations.SerializedName
 import java.util.*
 
@@ -27,4 +28,15 @@ data class PictureOfTheDayDTO(
 
     @SerializedName("url")
     val url: String
+)
+
+fun PictureOfTheDayDTO.toPictureOfTheDay(): PictureOfTheDay = PictureOfTheDay(
+    copyright = copyright,
+    date = date,
+    explanation = explanation,
+    hdurl = hdurl,
+    mediaType = mediaType,
+    serviceVersion = serviceVersion,
+    title = title,
+    url = url
 )
