+++ app/src/main/java/com/example/npod/ui/viewmodels/MainViewModel.kt
@@ -1,10 +1,9 @@
 package com.example.npod.ui.viewmodels
 
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
 import com.example.npod.data.PictureState
-import com.example.npod.domain.repository.NasaRepository
+import com.example.npod.domain.NasaRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.launch
